requirements.txt
# python
streamlit==1.34.0
boto3==1.34.98
biopython==1.85
# datasketch
# re
# pyspark
omicverse==1.6.10
pandas==2.2.2
requests